+++ CHANGELOG.md
@@ -8,6 +8,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 
 ### Added
 
+- Docker registry credentials honeytoken type for `~/.docker/config.json`.
+
 ### Changed
 
 ### Fixed